Fix #4 - refactoring code
DirectlineClient 코드에서 메세지 포멧팅 및 파싱하는 부분의 코드를 utils 폴더에 export 하였습니다.
