+++ utils/parseArray.js
@@ -1,0 +1,8 @@
+function parseArray (arr) {
+  var temp = ''
+  arr.forEach(val => {
+    temp += val + '\n'
+  })
+  return temp.slice(0, -1)
+}
+module.exports = parseArray
